Turn off race detection for upload test
