+++ .github/workflows/go-ci.yml
@@ -17,4 +17,6 @@ jobs:
         uses: actions/setup-go@v5
         with:
           go-version: ${{ matrix.go-version }}
-      - run: go test -race -v ./...
+      - run: |
+        go test -race -v ./...
+        go test -v ./...
